[story6][Teddy]refactor:convert parkingLot to List type

+++ src/main/java/com/oocl/cultivation/AvailableRateComparator.java
@@ -6,7 +6,7 @@ public class AvailableRateComparator implements Comparator<ParkingLot> {
 
     @Override
     public int compare(ParkingLot o1, ParkingLot o2) {
-        if ((o1.getCapacity() - o1.getParkingLot().size()) / o1.getCapacity() < (o2.getCapacity() - o2.getParkingLot().size()) / o2.getCapacity()) {
+        if ((o1.getCapacity() - o1.getParkingLot().size()) / (double)o1.getCapacity() < (o2.getCapacity() - o2.getParkingLot().size()) / (double)o2.getCapacity()) {
             return 1;
         } else {
             return -1;
